+++ packages/in_app_messaging/lib/src/data/model/conditions/once_per_event_properties_condition.dart
@@ -3,18 +3,32 @@ import 'package:in_app_messaging/src/domain/entity/message_condition.dart';
 class OncePerEventPropertiesCondition extends MessageCondition {
   final Set<String> properties;
 
-  OncePerEventPropertiesCondition({required this.properties});
+  /// if [distinct] is true, it will compare properties of event triggered just now with previous events of same event type
+  ///
+  /// otherwise it will compare properties of event triggered just now with all previous events.
+  final bool distinct;
+
+  OncePerEventPropertiesCondition({
+    required this.properties,
+    this.distinct = true,
+  });
 
   @override
   dynamic asJsonLogic() {
     return {
       "!": {
         "in": [
-          properties.map((e) => {"var": "event.$e"}).toList(),
+          [
+            ...properties.map((e) => {"var": "event.$e"}),
+            if (distinct) {"var": "event.event_name"}
+          ],
           {
             "map": [
               {"var": "interactions.seen_entries"},
-              properties.map((e) => {"var": "trigger_properties.$e"}).toList(),
+              [
+                ...properties.map((e) => {"var": "trigger_properties.$e"}),
+                if (distinct) {"var": "trigger"}
+              ],
             ]
           },
         ]