+++ packages/in_app_messaging/lib/src/data/model/conditions/once_per_time_condition.dart
@@ -25,7 +25,7 @@ class OncePerTimeCondition extends MessageCondition {
                 'microseconds'
               ]
             },
-            {'var': 'interactions.last_seen'}
+            {'var': 'interactions.last_seen.date'}
           ]
         }
       ]