+++ packages/in_app_messaging/test/conditions_test.dart
@@ -1,5 +1,6 @@
 import 'package:in_app_messaging/in_app_messaging.dart';
 import 'package:in_app_messaging/src/core/typedefs.dart';
+import 'package:in_app_messaging/src/data/data.dart';
 import 'package:in_app_messaging_json_logic/in_app_messaging_json_logic.dart';
 import 'package:test/test.dart';
 
@@ -17,6 +18,52 @@ final today = DateTime.now();
 final yesterday = today.subtract(const Duration(days: 1));
 
 final cases = [
+  // Once
+  _TestCase(
+    'once - never seen',
+    OnceCondition().asJsonLogic(),
+    {
+      'interactions': {'last_seen': never}
+    },
+    true,
+  ),
+  _TestCase(
+    'once - seen today',
+    OncePerDayCondition().asJsonLogic(),
+    {
+      'interactions': {
+        'last_seen': {"date": today}
+      }
+    },
+    false,
+  ),
+  _TestCase(
+    'once - never seen',
+    OnceCondition().asJsonLogic(),
+    {'interactions': {}},
+    true,
+  ),
+  _TestCase(
+    'once - seen',
+    OnceCondition().asJsonLogic(),
+    {
+      'interactions': {
+        'last_seen': {"date": yesterday}
+      }
+    },
+    false,
+  ),
+  _TestCase(
+    'once - seen in far future',
+    OnceCondition().asJsonLogic(),
+    {
+      'interactions': {
+        'last_seen': {"date": today.add(const Duration(days: 1000))}
+      }
+    },
+    false,
+  ),
+
   // Once per day
   _TestCase(
     'once per day - never seen',
@@ -30,7 +77,9 @@ final cases = [
     'once per day - seen today',
     OncePerDayCondition().asJsonLogic(),
     {
-      'interactions': {'last_seen': today}
+      'interactions': {
+        'last_seen': {"date": today}
+      }
     },
     false,
   ),
@@ -38,7 +87,9 @@ final cases = [
     'once per day - seen yesterday',
     OncePerDayCondition().asJsonLogic(),
     {
-      'interactions': {'last_seen': yesterday}
+      'interactions': {
+        'last_seen': {"date": yesterday}
+      }
     },
     true,
   ),
@@ -56,7 +107,9 @@ final cases = [
     'once per hour - seen 55 minutes ago',
     OncePerTimeCondition(duration: const Duration(hours: 1)).asJsonLogic(),
     {
-      'interactions': {'last_seen': today.subtract(const Duration(minutes: 55))}
+      'interactions': {
+        'last_seen': {"date": today.subtract(const Duration(minutes: 55))}
+      }
     },
     false,
   ),
@@ -64,7 +117,9 @@ final cases = [
     'once per hour - seen 2 hours ago',
     OncePerTimeCondition(duration: const Duration(hours: 1)).asJsonLogic(),
     {
-      'interactions': {'last_seen': today.subtract(const Duration(hours: 2))}
+      'interactions': {
+        'last_seen': {"date": today.subtract(const Duration(hours: 2))}
+      }
     },
     true,
   ),
@@ -72,31 +127,9 @@ final cases = [
     'once per hour - seen 2 hours in future',
     OncePerTimeCondition(duration: const Duration(hours: 1)).asJsonLogic(),
     {
-      'interactions': {'last_seen': today.add(const Duration(hours: 2))}
-    },
-    false,
-  ),
-
-  // Once condition
-  _TestCase(
-    'once - never seen',
-    OnceCondition().asJsonLogic(),
-    {'interactions': {}},
-    true,
-  ),
-  _TestCase(
-    'once - seen',
-    OnceCondition().asJsonLogic(),
-    {
-      'interactions': {'last_seen': yesterday}
-    },
-    false,
-  ),
-  _TestCase(
-    'once - seen in far future',
-    OnceCondition().asJsonLogic(),
-    {
-      'interactions': {'last_seen': today.add(const Duration(days: 1000))}
+      'interactions': {
+        'last_seen': {"date": today.add(const Duration(hours: 2))}
+      }
     },
     false,
   ),
@@ -106,7 +139,7 @@ final cases = [
     'once per trigger properties - never seen',
     OncePerEventPropertiesCondition(properties: {'id', 'name'}).asJsonLogic(),
     {
-      'event': {'id': '1', 'name': 'John'},
+      'event': {'event_name': 'event_1', 'id': '1', 'name': 'John'},
       'interactions': {'last_seen': never, 'seen_entries': []}
     },
     true,
@@ -115,12 +148,13 @@ final cases = [
     'once per trigger properties - seen with other trigger properties',
     OncePerEventPropertiesCondition(properties: {'id', 'name'}).asJsonLogic(),
     {
-      'event': {'id': '1', 'name': 'John'},
+      'event': {'event_name': 'event_1', 'id': '1', 'name': 'John'},
       'interactions': {
-        'last_seen': yesterday,
+        'last_seen': {"date": yesterday},
         'seen_entries': [
           {
             'date': yesterday,
+            'trigger': 'event_1',
             'trigger_properties': {'id': '2', 'name': 'Jane'}
           }
         ]
@@ -132,14 +166,139 @@ final cases = [
     'once per trigger properties - seen with same trigger properties',
     OncePerEventPropertiesCondition(properties: {'id', 'name'}).asJsonLogic(),
     {
-      'event': {'id': '1', 'name': 'John'},
+      'event': {'event_name': 'event_1', 'id': '1', 'name': 'John'},
       'interactions': {
-        'last_seen': yesterday,
+        'last_seen': {"date": yesterday},
         'seen_entries': [
           {
             'date': yesterday,
+            'trigger': 'event_1',
             'trigger_properties': {'id': '1', 'name': 'John'}
           }
+        ]
+      }
+    },
+    false,
+  ),
+  _TestCase(
+    'once per trigger properties - other event - seen with other trigger properties',
+    OncePerEventPropertiesCondition(properties: {'id', 'name'}).asJsonLogic(),
+    {
+      'event': {'event_name': 'event_1', 'id': '1', 'name': 'John'},
+      'interactions': {
+        'last_seen': {"date": yesterday},
+        'seen_entries': [
+          {
+            'date': yesterday,
+            'trigger': 'event_2',
+            'trigger_properties': {'id': '2', 'name': 'Jane'}
+          }
+        ]
+      }
+    },
+    true,
+  ),
+  _TestCase(
+    'once per trigger properties - other event - seen with same trigger properties',
+    OncePerEventPropertiesCondition(properties: {'id', 'name'}).asJsonLogic(),
+    {
+      'event': {'event_name': 'event_1', 'id': '1', 'name': 'John'},
+      'interactions': {
+        'last_seen': {"date": yesterday},
+        'seen_entries': [
+          {
+            'date': yesterday,
+            'trigger': 'event_2',
+            'trigger_properties': {'id': '1', 'name': 'John'}
+          }
+        ]
+      }
+    },
+    true,
+  ),
+  _TestCase(
+    'once per trigger properties - other event - distinct - seen with same trigger properties',
+    OncePerEventPropertiesCondition(properties: {'id', 'name'}, distinct: false)
+        .asJsonLogic(),
+    {
+      'event': {'event_name': 'event_1', 'id': '1', 'name': 'John'},
+      'interactions': {
+        'last_seen': {"date": yesterday},
+        'seen_entries': [
+          {
+            'date': yesterday,
+            'trigger': 'event_2',
+            'trigger_properties': {'id': '1', 'name': 'John'}
+          }
+        ]
+      }
+    },
+    false,
+  ),
+  _TestCase(
+    'once per trigger properties - other event, this event - seen with other trigger properties',
+    OncePerEventPropertiesCondition(properties: {'id', 'name'}).asJsonLogic(),
+    {
+      'event': {'event_name': 'event_1', 'id': '1', 'name': 'John'},
+      'interactions': {
+        'last_seen': {"date": yesterday},
+        'seen_entries': [
+          {
+            'date': yesterday,
+            'trigger': 'event_2',
+            'trigger_properties': {'id': '2', 'name': 'Jane'}
+          },
+          {
+            'date': yesterday,
+            'trigger': 'event_1',
+            'trigger_properties': {'id': '2', 'name': 'Jane'}
+          }
+        ]
+      }
+    },
+    true,
+  ),
+  _TestCase(
+    'once per trigger properties - other event, this event - seen with same trigger properties',
+    OncePerEventPropertiesCondition(properties: {'id', 'name'}).asJsonLogic(),
+    {
+      'event': {'event_name': 'event_1', 'id': '1', 'name': 'John'},
+      'interactions': {
+        'last_seen': {"date": yesterday},
+        'seen_entries': [
+          {
+            'date': yesterday,
+            'trigger': 'event_2',
+            'trigger_properties': {'id': '1', 'name': 'John'}
+          },
+          {
+            'date': yesterday,
+            'trigger': 'event_1',
+            'trigger_properties': {'id': '1', 'name': 'John'}
+          }
+        ]
+      }
+    },
+    false,
+  ),
+  _TestCase(
+    'once per trigger properties - other event, this event - reversed - seen with same trigger properties',
+    OncePerEventPropertiesCondition(properties: {'id', 'name'}).asJsonLogic(),
+    {
+      'event': {'event_name': 'event_1', 'id': '1', 'name': 'John'},
+      'interactions': {
+        'last_seen': {"date": yesterday},
+        'seen_entries': [
+          {
+            'date': yesterday,
+            'trigger': 'event_1',
+            'trigger_properties': {'id': '1', 'name': 'John'}
+          },
+          {
+            'date': yesterday,
+            'trigger': 'event_2',
+            'trigger_properties': {'id': '1', 'name': 'John'}
+          },
         ]
       }
     },